+++ tests/testthat/test_taxaid_RedLake.R
@@ -42,7 +42,33 @@ test_that("RedLake, trans, dups1", {
   
 })## Test ~ dup, trans
 
-
+test_that("RedLake, trans, dups2", {
+  # data files
+  df_pickfiles <- read.csv(file.path("data"
+                                     , "taxa_official"
+                                     , "_pick_files_MN.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  ii <- 2# + 4
+  
+  # for (i in seq_len(length(fn_all))) {
+  fn_ii <- df_pickfiles[ii , "filename"]
+  df_ii <- read.csv(file.path("data", "taxa_official", fn_ii))
+  ii_taxaid <- df_pickfiles[ii, "taxaid"]
+  
+  # QC
+  n_taxa_calc <- length(unique(df_ii[, ii_taxaid]))
+  n_taxa_qc   <- length(df_ii[, ii_taxaid])
+  
+  # show mismatches
+  print(paste0("Duplicate names '", ii_taxaid, "' from file '", fn_ii, "'"))
+  n_taxa_table <- as.data.frame(table(df_ii[, ii_taxaid]))
+  n_taxa_table[n_taxa_table$Freq > 1, ]
+  
+  #test
+  testthat::expect_equal(n_taxa_calc, n_taxa_qc)
+  
+})## Test ~ 2 dup, trans
 
 test_that("RedLake, attr, dups1", {
   
@@ -73,7 +99,34 @@ test_that("RedLake, attr, dups1", {
   
 })## Test ~ dup, trans
 
-
+test_that("RedLake, attr, dups2", {
+  
+  # data files
+  df_pickfiles <- read.csv(file.path("data"
+                                     , "taxa_official"
+                                     , "_pick_files_MN.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  jj <- 2 #+ 6
+  
+  # attribute file
+  fn_jj <- df_pickfiles[jj , "attributes_filename"]
+  df_jj <- read.csv(file.path("data", "taxa_official", fn_jj))
+  jj_taxaid <- df_pickfiles[jj, "attributes_taxaid"]
+  
+  # QC
+  n_taxa_calc <- length(unique(df_jj[, jj_taxaid]))
+  n_taxa_qc   <- length(df_jj[, jj_taxaid])
+  
+  # show mismatches
+  print(paste0("Duplicate names '", jj_taxaid, "' from file '", fn_jj, "'"))
+  n_taxa_table <- as.data.frame(table(df_jj[, jj_taxaid]))
+  n_taxa_table[n_taxa_table$Freq > 1, ]
+  
+  # test
+  testthat::expect_equivalent(n_taxa_calc, n_taxa_qc)
+  
+})## Test ~ dup, trans
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ## trans, taxaid ----
@@ -104,7 +157,7 @@ test_that("RedLake, trans, taxaid_1", {
     sort(i1)
     
     # Remove DNI
-    n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
+    #n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
     
     # test, match
     testthat::expect_equivalent(n_match_calc, n_match_QC)
@@ -118,6 +171,50 @@ test_that("RedLake, trans, taxaid_1", {
     testthat::expect_equivalent(n_rows, n_taxaid)
 
  # }## FOR ~ i
+  
+})## Test ~ taxaid
+
+test_that("RedLake, trans, taxaid_2", {
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_MN.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  i <- 2# + 6
+  
+  # for (i in seq_len(length(fn_all))) {
+  fn_i <- df_pickfiles[i , "filename"]
+  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
+  i_taxaid <- df_pickfiles[i, "taxaid"]
+  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
+  
+  # # Remove DNI
+  # df_i <- df_i[!df_i[, i_taxaid_match] == "DNI", ]
+  
+  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_i[, i_taxaid])
+  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
+  
+  
+  # show mismatches
+  print(paste0("Unique '", i_taxaid_match, "' missing from '", i_taxaid, "'"))
+  i1 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
+                                         df_i[, i_taxaid]]
+  sort(i1)
+  
+  # Remove DNI
+  #n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
+  
+  # test, match
+  testthat::expect_equivalent(n_match_calc, n_match_QC)
+  
+  # test, dup
+  n_taxaid <- sum(length(unique(df_i[, i_taxaid])))
+  n_rows <- nrow(df_i)
+  df_freq <- as.data.frame(table(df_i[, i_taxaid]))
+  df_freq[df_freq$Freq > 1, "Var1"]
+  
+  testthat::expect_equivalent(n_rows, n_taxaid)
+  
+  # }## FOR ~ i
   
 })## Test ~ taxaid
     
@@ -163,7 +260,7 @@ test_that("RedLake, attr, taxaid1", {
   sort(j1)
   
   # Remove DNI
-  n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
+  #n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
   
   # test
   testthat::expect_equivalent(n_match_calc, n_match_QC)
@@ -181,6 +278,51 @@ test_that("RedLake, attr, taxaid1", {
   
 })## Test ~ taxaid
 
+test_that("RedLake, attr, taxaid2", {
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_MN.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  i <- 2# + 6
+  
+  # for (i in seq_len(length(fn_all))) {
+  fn_i <- df_pickfiles[i , "filename"]
+  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
+  i_taxaid <- df_pickfiles[i, "taxaid"]
+  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
+  
+  # attribute file
+  fn_j <- df_pickfiles[i , "attributes_filename"]
+  df_j <- read.csv(file.path("data", "taxa_official", fn_j))
+  j_taxaid <- df_pickfiles[i, "attributes_taxaid"]
+  
+  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_j[, j_taxaid])
+  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
+  
+  # show mismatches
+  print(paste0("Unique '", j_taxaid, "' missing from '", i_taxaid_match, "'"))
+  j1 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
+                                         df_j[, j_taxaid]]
+  sort(j1)
+  
+  # Remove DNI
+  #n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
+  
+  # test
+  testthat::expect_equivalent(n_match_calc, n_match_QC)
+  
+  # test, dup
+  n_taxaid <- sum(length(unique(df_j[, j_taxaid])))
+  n_rows <- nrow(df_j)
+  df_freq <- as.data.frame(table(df_j[, j_taxaid]))
+  df_freq[df_freq$Freq > 1, "Var1"]
+  
+  testthat::expect_equivalent(n_rows, n_taxaid)
+  
+  
+  # }## FOR ~ i
+  
+})## Test ~ taxaid
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -214,6 +356,31 @@ test_that("RedLake, trans, dups3", {
   
 })## Test ~ dup, trans
 
+test_that("RedLake, trans, dups4", {
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_MN.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  ii <- 4# + 4
+  
+  # for (i in seq_len(length(fn_all))) {
+  fn_ii <- df_pickfiles[ii , "filename"]
+  df_ii <- read.csv(file.path("data", "taxa_official", fn_ii))
+  ii_taxaid <- df_pickfiles[ii, "taxaid"]
+  
+  # QC
+  n_taxa_calc <- length(unique(df_ii[, ii_taxaid]))
+  n_taxa_qc   <- length(df_ii[, ii_taxaid])
+  
+  # show mismatches
+  print(paste0("Duplicate names '", ii_taxaid, "' from file '", fn_ii, "'"))
+  n_taxa_table <- as.data.frame(table(df_ii[, ii_taxaid]))
+  n_taxa_table[n_taxa_table$Freq > 1, ]
+  
+  #test
+  testthat::expect_equal(n_taxa_calc, n_taxa_qc)
+  
+})## Test ~ dup, trans
 
 test_that("RedLake, attr, dups3", {
   
@@ -222,6 +389,34 @@ test_that("RedLake, attr, dups3", {
   fn_all <- df_pickfiles$filename
   
   jj <- 3 #+ 6
+  
+  # attribute file
+  fn_jj <- df_pickfiles[jj , "attributes_filename"]
+  df_jj <- read.csv(file.path("data", "taxa_official", fn_jj))
+  jj_taxaid <- df_pickfiles[jj, "attributes_taxaid"]
+  
+  # QC
+  n_taxa_calc <- length(unique(df_jj[, jj_taxaid]))
+  n_taxa_qc   <- length(df_jj[, jj_taxaid])
+  
+  # show mismatches
+  print(paste0("Duplicate names '", jj_taxaid, "' from file '", fn_jj, "'"))
+  n_taxa_table <- as.data.frame(table(df_jj[, jj_taxaid]))
+  n_taxa_table[n_taxa_table$Freq > 1, ]
+  
+  # test
+  testthat::expect_equivalent(n_taxa_calc, n_taxa_qc)
+  
+})## Test ~ dup, trans
+
+
+test_that("RedLake, attr, dups4", {
+  
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_MN.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  jj <- 4 #+ 6
   
   # attribute file
   fn_jj <- df_pickfiles[jj , "attributes_filename"]
@@ -277,9 +472,40 @@ test_that("RedLake, trans, taxaid_3", {
   
   # }## FOR ~ i
   
-})## Test ~ taxaid 2
+})## Test ~ taxaid 3
 
 
+test_that("RedLake, trans, taxaid_4", {
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_MN.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  i <- 4# + 6
+  
+  # for (i in seq_len(length(fn_all))) {
+  fn_i <- df_pickfiles[i , "filename"]
+  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
+  i_taxaid <- df_pickfiles[i, "taxaid"]
+  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
+  
+  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_i[, i_taxaid])
+  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
+  
+  # show mismatches
+  print(paste0("Unique '", i_taxaid_match, "' missing from '", i_taxaid, "'"))
+  i3 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
+                                         df_i[, i_taxaid]]
+  sort(i3)
+  
+  # Remove DNI
+  n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
+  
+  # test
+  testthat::expect_equivalent(n_match_calc, n_match_QC)
+  
+  # }## FOR ~ i
+  
+})## Test ~ taxaid 4
 
 
 
@@ -327,7 +553,44 @@ test_that("RedLake, attr, taxaid_3", {
   
   # }## FOR ~ i
   
-})## Test ~ taxaid 3
+})## Test ~ taxaid 4
+
+test_that("RedLake, attr, taxaid_3", {
+  # data files
+  df_pickfiles <- read.csv(file.path("data", "taxa_official", "_pick_files_MN.csv"))
+  fn_all <- df_pickfiles$filename
+  
+  i <- 4 #+ 6
+  
+  # for (i in seq_len(length(fn_all))) {
+  fn_i <- df_pickfiles[i , "filename"]
+  df_i <- read.csv(file.path("data", "taxa_official", fn_i))
+  i_taxaid <- df_pickfiles[i, "taxaid"]
+  i_taxaid_match <- df_pickfiles[i, "calc_taxaid"]
+  
+  # attribute file
+  fn_j <- df_pickfiles[i , "attributes_filename"]
+  df_j <- read.csv(file.path("data", "taxa_official", fn_j))
+  j_taxaid <- df_pickfiles[i, "attributes_taxaid"]
+  
+  n_match_calc <- sum(unique(df_i[, i_taxaid_match]) %in% df_j[, j_taxaid])
+  n_match_QC <- length(unique(df_i[, i_taxaid_match]))
+  
+  # show mismatches
+  print(paste0("Unique '", j_taxaid, "' missing from '", i_taxaid_match, "'"))
+  j3 <- unique(df_i[, i_taxaid_match])[!unique(df_i[, i_taxaid_match]) %in% 
+                                         df_j[, j_taxaid]]
+  sort(j3)
+  
+  # Remove DNI
+  n_match_QC <- n_match_QC - "DNI" %in% df_i[, i_taxaid_match]
+  
+  # test
+  testthat::expect_equivalent(n_match_calc, n_match_QC)
+  
+  # }## FOR ~ i
+  
+})## Test ~ taxaid 4
 
 # OLD ----
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~